requirements.txt
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap4==4.0.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
python-dotenv==0.10.5
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
gunicorn==20.0.4  # for heroku